Add Project DbFirstVidzyOracle 1st Iteration
The projejct DbFirstVidzyOracle is a practice implementing Oracle Database (Express Edition).
Before create the model its necesary to install Oracle Entity Framework, for this project i use the next command:

Install-Package Oracle.ManagedDataAccess.EntityFramework -Version 12.2.1100

After this the model from the database can be created.

In the app.config i comment some line what refer to EF with SQL Server that are not necesary for Oracle.

In the connection string to the EF Model I use a Data Source that exists in the tnanames.ora, so there is no need to configure the datasource in configuration file.