Workaround parameter Classification EnumType
One way to solver the problem with Parameter is edit the Store Schema manually and change the precison (in this case to 3) that way EF maps the parameter to byte.

The problem with this workaround is that every time the is model is updated from the database this changes are lost.