Enable Migrations Oracle
The project is enable to work with Entity Framework Code First with command Enable-Migrations.

To work with oracle is needed to define the connection string in the config file and reference that connection in the DbContext class even when the name of the DbContext is the same that the connection string.

In this connection string the DataSource work using the configuration in the tnsname.ora file.