2nd Iteration CodeFirstOracle
The relationship between Videos and Genres change from One-to-Many to One-to-One.
This migration delete the VideoGenres table and add a column GenreId in Videos.

In the exercise with SQL Server substitute the property IList<Genre> Genres in the class Video with single Genre was enough for understand the relationship change.

In Oracle it was necesary add a property class GenreId, because without this property EF was renaming the table VideoGenres to Video.