FluentAPIVidzyOracle 4th Iteration Fix
For applying the last migration Oracle not support the alter column datatype nclob to nvarchar2/varchar2.

It was necesary rename the column like "_old", add new column but like varchar2 (maxLength: 3999, unicode: false), migrate data to new column and delete the "_old" then make the alter table generate by the add-migration command.

EF and Oracle in FluentAPI require that if you want a column to be VARCHAR2 the unicode need to be false.

Check documentation for Mapping Method and DataTypes:
https://docs.oracle.com/cd/E63277_01/win.121/e63268/entityCodeFirst.htm#ODPNT8310